Create a FAQ based on email exchanges.
It's not exactly what you'd want in a FAQ, but it aggregates most of my
email exchanges with Tony and will hopefully be useful when creating a
more comprehensive collection of documentation.

I made a small effort to edit some of the information to be more usable
to a newcomer, but a lot of it is missing important context (such as
having read the paper describing the Network Calculus).