Use host-route netmask in gateway outbound relay.
I'm not sure why previously it had been hardcoded to a 32-bit netmask;
presumably this was an error on my part way back in the original routing
implementation in minimart-netstack. It looks as if the code was
originally written for a kind of host route that didn't have a netmask,
and was never updated to include the netmask later.