Track fragment versions to reestablish child fragments after a change.
Without this, if I edit, say, post-entry.html, then the actual rendered
post items disappear, because the UI protocol isn't sufficient to allow
replacement of a parent location with survival of a child. I am not sure
if this is a "fix" or a "workaround": it's workaroundish in that in
principle a child fragment could monitor its parent fragment and rebuild
itself when it detects a change. I suspect revision to the UI protocols
is indicated.